+++ packages/state-manager/src/sagas/publicChannels/markUnreadChannels/markUnreadChannels.saga.test.ts
@@ -144,9 +144,8 @@ describe('markUnreadChannelsSaga', () => {
     const messagesides = channelIds
     const messages: ChannelMessage[] = []
 
-    const community = await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>(
-      'Community'
-    )
+    const community =
+      await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>('Community')
 
     const alice = await factory.create<ReturnType<typeof identityActions.addNewIdentity>['payload']>('Identity', {
       id: community.id,