+++ packages/state-manager/src/sagas/publicChannels/sendNewUserInfoMessage/sendNewUserInfoMessage.saga.test.ts
@@ -34,9 +34,8 @@ describe('sendInitialChannelMessageSaga', () => {
   })
 
   test('send new user info message', async () => {
-    const community1 = await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>(
-      'Community'
-    )
+    const community1 =
+      await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>('Community')
     user = await factory.create<ReturnType<typeof identityActions.addNewIdentity>['payload']>('Identity', {
       id: community1.id,
     })
@@ -78,9 +77,8 @@ describe('sendInitialChannelMessageSaga', () => {
   })
 
   test('dont send new user info if user exists', async () => {
-    const community1 = await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>(
-      'Community'
-    )
+    const community1 =
+      await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>('Community')
     user = await factory.create<ReturnType<typeof identityActions.addNewIdentity>['payload']>('Identity', {
       id: community1.id,
     })
@@ -117,9 +115,8 @@ describe('sendInitialChannelMessageSaga', () => {
   })
 
   test('dont send new user info message if owner', async () => {
-    const community1 = await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>(
-      'Community'
-    )
+    const community1 =
+      await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>('Community')
     user = await factory.create<ReturnType<typeof identityActions.addNewIdentity>['payload']>('Identity', {
       id: community1.id,
     })
@@ -145,9 +142,8 @@ describe('sendInitialChannelMessageSaga', () => {
   })
 
   test('remove possible duplicates before sending info message', async () => {
-    const community1 = await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>(
-      'Community'
-    )
+    const community1 =
+      await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>('Community')
     user = await factory.create<ReturnType<typeof identityActions.addNewIdentity>['payload']>('Identity', {
       id: community1.id,
     })